+++ internal/database/model/model.go
@@ -1,4 +1,4 @@
-package database
+package model
 
 import (
 	"time"
