+++ scripts/migrate.go
@@ -1,14 +1,19 @@
 package main
 
-import "github.com/Misoten-B/airship-backend/internal/database"
+import (
+	"github.com/Misoten-B/airship-backend/internal/database"
+	"github.com/Misoten-B/airship-backend/internal/database/model"
+)
 
 func main() {
+	// データベースに接続
 	db, err := database.ConnectDB()
 	if err != nil {
 		panic(err)
 	}
 
-	if err = db.AutoMigrate(&database.User{}); err != nil {
+	// マイグレーション
+	if err = db.AutoMigrate(&model.User{}); err != nil {
 		panic(err)
 	}
 }
